requirements.txt
asgiref==3.8.1
decorator==5.1.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
self==2020.12.3
sqlparse==0.5.0
typing_extensions==4.11.0
